[feat] redis 관련 설정 추가, products 모델 변경, api 수정

+++ likeProducts/models.py
@@ -1,6 +1,5 @@
 from django.db import models
 from accounts.models import Account
-from products.models import Product
 
 
 # Create your models here.
@@ -8,8 +7,8 @@ class LikeProduct(models.Model):
   id = models.AutoField(primary_key=True)
   user = models.ForeignKey(Account, on_delete=models.CASCADE)
   name = models.CharField(max_length=100)
-  price = models.DecimalField(max_digits=10, decimal_places=2)
-  delivery_charge = models.DecimalField(max_digits=10, decimal_places=2, default=0)
+  price = models.IntegerField()
+  delivery_charge = models.IntegerField(default=0)
   link = models.URLField()
   image_url = models.URLField()
   category_id = models.IntegerField()
